Made script serialize

+++ bitcoin/script/script.h
@@ -8,19 +8,19 @@
 
 typedef struct {
     unsigned char data[CMDS_SIZE];
-    size_t data_len;
+    int data_len;
 } Command;
 
 typedef struct {
     Command cmds[CMDS_SIZE];
-    size_t cmds_len;
+    int cmds_len;
 } Script;
 
 Script* script_init();
 
 void script_free(Script* script);
 
-void script_set_cmds(Script* script, Command* cmds, size_t cmds_len);
+void script_set_cmds(Script* script, Command* cmds, int cmds_len);
 
 Script* script_parse(unsigned char* s);
